backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
reportlab==4.0.7
requests==2.31.0
cryptography==41.0.7
